+++ tests/test_search.py
@@ -52,7 +52,7 @@ def test_all_objects(httpserver):
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SyncSCIMClient(
         client,
-        resource_types=(
+        resource_models=(
             User,
             Group,
         ),
@@ -129,7 +129,7 @@ def test_search_request(httpserver):
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SyncSCIMClient(
         client,
-        resource_types=(
+        resource_models=(
             User,
             Group,
         ),
@@ -158,7 +158,7 @@ def test_dont_check_response(httpserver):
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SyncSCIMClient(
         client,
-        resource_types=(
+        resource_models=(
             User,
             Group,
         ),
@@ -205,7 +205,7 @@ def test_dont_check_request_payload(httpserver):
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SyncSCIMClient(
         client,
-        resource_types=(
+        resource_models=(
             User,
             Group,
         ),
@@ -229,7 +229,7 @@ def test_errors(httpserver, code):
     client = Client(base_url=f"http://localhost:{httpserver.port}")
     scim_client = SyncSCIMClient(
         client,
-        resource_types=(
+        resource_models=(
             User,
             Group,
         ),
@@ -246,7 +246,7 @@ def test_errors(httpserver, code):
 def test_request_network_error(httpserver):
     """Test that httpx exceptions are transformed in RequestNetworkError."""
     client = Client(base_url=f"http://localhost:{httpserver.port}")
-    scim_client = SyncSCIMClient(client, resource_types=(User,))
+    scim_client = SyncSCIMClient(client, resource_models=(User,))
     with pytest.raises(
         RequestNetworkError, match="Network error happened during request"
     ):